Deleted unused PlannerClasses
