+++ app/src/main/java/com/android/sample/ui/planner/PlannerTestTags.kt
@@ -1,0 +1,14 @@
+package com.android.sample.ui.planner
+
+object PlannerScreenTestTags {
+  const val PLANNER_SCREEN = "plannerScreen"
+  const val PET_HEADER = "petHeader" // Renamed from PET_SECTION to avoid conflict
+  const val TODAY_CLASSES_SECTION = "TODAY_CLASSES_SECTION" // Renamed for broader scope
+  const val WELLNESS_CAMPUS_SECTION = "WELLNESS_CAMPUS_SECTION" // New section
+  const val ADD_TASK_MODAL = "addTaskModal"
+  const val CLASS_ATTENDANCE_MODAL = "classAttendanceModal"
+  const val SUBJECT_FIELD = "subject_field"
+  const val TASK_TITLE_FIELD = "task_title_field"
+  const val DURATION_FIELD = "duration_field"
+  const val DEADLINE_FIELD = "deadline_field"
+}
